refactor(catalogue): extract the series layer to clear the source ceiling
CI red on one root cause reported twice: `check_source_size_budget` (Rule
Backstops) and its own test, both on 19615 against a baseline of 19611.
The R2 fixes pushed `skill_catalogue.ts` to 1,504 lines, four past the
1,500 ceiling.

Extracted, not baselined -- the ratchet's own doctrine, and the note in
`gate-violation-baselines.json` states the ordering: extract, then measure.
Raising a baseline to clear a failing check is the config-weakening move
this repo blocks by construction.

The seam is real rather than a convenient cut at line 1500.
`skill_catalogue` answers what one observation measured; the new
`skill_catalogue_series` answers the three questions that need more than
one record -- is a host's series stale, how far is the corpus from its bar,
and does host delivery diverge from the disk the ranker reads. Nothing in
it is imported by `install.ts`, so the deploy-time warning path keeps
exactly the surface it had, and the committed bundle is unchanged
(verified by rebuild, not assumed).

The CLI re-exports both halves, so the split is an internal seam and not a
second import path consumers have to learn -- no test or caller changed.

19611 total excess, back at baseline. 63 tests green across the three
affected files.
